rename to chaintest, slice of `ActionTest`, remove suite
