+++ examples/tokenvm/storage/storage.go
@@ -19,6 +19,7 @@ import (
 	"github.com/ava-labs/hypersdk/state"
 
 	smath "github.com/ava-labs/avalanchego/utils/math"
+
 	tconsts "github.com/ava-labs/hypersdk/examples/tokenvm/consts"
 )
